# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  profileImageUrl?: string | null;
}

export interface GetCurrentAuthUserResponse {
  user: AuthUser | null;
}

export interface ExchangeMobileAuthorizationCodeBody {
  code: string;
  code_verifier: string;
  redirect_uri: string;
  state: string;
  /** @nullable */
  nonce?: string | null;
}

export interface ExchangeMobileAuthorizationCodeResponse {
  token: string;
}

export interface LogoutMobileSessionResponse {
  success: boolean;
}

export interface UserProfile {
  userId: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  shabbosCity?: string | null;
  onboardingComplete: boolean;
  shareCode: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  hobbies?: string | null;
  /** @nullable */
  growthGoals?: string | null;
}

export interface UpsertProfileBody {
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  shabbosCity?: string | null;
  onboardingComplete?: boolean;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  hobbies?: string | null;
  /** @nullable */
  growthGoals?: string | null;
}

export interface FriendEntry {
  userId: string;
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  profileImageUrl?: string | null;
  shareCode: string;
}

export interface AddFriendBody {
  shareCode: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface VideoEntry {
  id: string;
  title: string;
  speaker: string;
  channel: string;
  duration: string;
  category: string;
  description: string;
  isShort: boolean;
  sortOrder: number;
  createdAt?: string;
}

export interface CreateVideoBody {
  id: string;
  title: string;
  speaker: string;
  channel: string;
  duration: string;
  category: string;
  description?: string;
  isShort?: boolean;
  sortOrder?: number;
}

export type ListVideosParams = {
  category?: string;
  isShort?: boolean;
};