# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get current auth user
 */
export const GetCurrentAuthUserResponse = zod.object({
  user: zod.union([
    zod.object({
      id: zod.string(),
      email: zod.string().nullish(),
      firstName: zod.string().nullish(),
      lastName: zod.string().nullish(),
      profileImageUrl: zod.string().nullish(),
    }),
    zod.null(),
  ]),
});

/**
 * @summary Exchange mobile auth code for session token
 */
export const ExchangeMobileAuthorizationCodeBody = zod.object({
  code: zod.string(),
  code_verifier: zod.string(),
  redirect_uri: zod.string(),
  state: zod.string(),
  nonce: zod.string().nullish(),
});

export const ExchangeMobileAuthorizationCodeResponse = zod.object({
  token: zod.string(),
});

/**
 * @summary Log out mobile session
 */
export const LogoutMobileSessionResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary Get current user profile
 */
export const GetProfileResponse = zod.object({
  userId: zod.string(),
  displayName: zod.string().nullish(),
  shabbosCity: zod.string().nullish(),
  onboardingComplete: zod.boolean(),
  shareCode: zod.string(),
  bio: zod.string().nullish(),
  hobbies: zod.string().nullish(),
  growthGoals: zod.string().nullish(),
});

/**
 * @summary Create or update user profile
 */
export const UpsertProfileBody = zod.object({
  displayName: zod.string().nullish(),
  shabbosCity: zod.string().nullish(),
  onboardingComplete: zod.boolean().optional(),
  bio: zod.string().nullish(),
  hobbies: zod.string().nullish(),
  growthGoals: zod.string().nullish(),
});

export const UpsertProfileResponse = zod.object({
  userId: zod.string(),
  displayName: zod.string().nullish(),
  shabbosCity: zod.string().nullish(),
  onboardingComplete: zod.boolean(),
  shareCode: zod.string(),
  bio: zod.string().nullish(),
  hobbies: zod.string().nullish(),
  growthGoals: zod.string().nullish(),
});

/**
 * @summary Get friend list for current user
 */
export const GetFriendsResponseItem = zod.object({
  userId: zod.string(),
  displayName: zod.string().nullish(),
  firstName: zod.string().nullish(),
  lastName: zod.string().nullish(),
  profileImageUrl: zod.string().nullish(),
  shareCode: zod.string(),
});
export const GetFriendsResponse = zod.array(GetFriendsResponseItem);

/**
 * @summary Add a friend by their share code
 */
export const AddFriendByCodeBody = zod.object({
  shareCode: zod.string(),
});

export const AddFriendByCodeResponse = zod.object({
  userId: zod.string(),
  displayName: zod.string().nullish(),
  firstName: zod.string().nullish(),
  lastName: zod.string().nullish(),
  profileImageUrl: zod.string().nullish(),
  shareCode: zod.string(),
});

/**
 * @summary Remove a friend
 */
export const RemoveFriendParams = zod.object({
  friendUserId: zod.coerce.string(),
});

export const RemoveFriendResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary List all videos
 */
export const ListVideosQueryParams = zod.object({
  category: zod.coerce.string().optional(),
  isShort: zod.coerce.boolean().optional(),
});

export const ListVideosResponseItem = zod.object({
  id: zod.string(),
  title: zod.string(),
  speaker: zod.string(),
  channel: zod.string(),
  duration: zod.string(),
  category: zod.string(),
  description: zod.string(),
  isShort: zod.boolean(),
  sortOrder: zod.number(),
  createdAt: zod.string().optional(),
});
export const ListVideosResponse = zod.array(ListVideosResponseItem);

/**
 * @summary Add or update a video (admin)
 */
export const CreateVideoBody = zod.object({
  id: zod.string(),
  title: zod.string(),
  speaker: zod.string(),
  channel: zod.string(),
  duration: zod.string(),
  category: zod.string(),
  description: zod.string().optional(),
  isShort: zod.boolean().optional(),
  sortOrder: zod.number().optional(),
});

export const CreateVideoResponse = zod.object({
  id: zod.string(),
  title: zod.string(),
  speaker: zod.string(),
  channel: zod.string(),
  duration: zod.string(),
  category: zod.string(),
  description: zod.string(),
  isShort: zod.boolean(),
  sortOrder: zod.number(),
  createdAt: zod.string().optional(),
});

/**
 * @summary Delete a video (admin)
 */
export const DeleteVideoParams = zod.object({
  videoId: zod.coerce.string(),
});

export const DeleteVideoResponse = zod.object({
  success: zod.boolean(),
});